+++ packages/native-core/lib/PDFRStreamForBuffer.js
@@ -25,11 +25,17 @@ PDFRStreamForBuffer.prototype.notEnded = function () {
 };
 
 PDFRStreamForBuffer.prototype.setPosition = function (inPosition) {
-  this.rposition = this.mStartPosition + inPosition;
+  this.rposition = Math.min(
+    Math.max(this.mStartPosition + inPosition, 0),
+    this.fileSize,
+  );
 };
 
 PDFRStreamForBuffer.prototype.setPositionFromEnd = function (inPosition) {
-  this.rposition = this.fileSize - inPosition;
+  this.rposition = Math.min(
+    Math.max(this.fileSize - inPosition, 0),
+    this.fileSize,
+  );
 };
 
 PDFRStreamForBuffer.prototype.skip = function (inAmount) {